requirements.txt
einops
librosa
numpy
omegaconf
pandas
pesq
ptflops
pystoi
soundfile
toml
torch_stoi
tqdm
